rag-agent/requirements.txt
langchain
langchain-community
langchain-anthropic
chromadb
pypdf
python-dotenv
anthropic
sentence-transformers

